实现功能：实现个人中心的小组管理 part2

+++ server/src/main/java/zheye/server/service/IGroupCollectService.java
@@ -23,5 +23,5 @@ public interface IGroupCollectService extends IService<GroupCollect> {
 
     int deleteGroupCollect(GroupCollect groupCollect);
 
-    List<GroupCollect> selectByGroupId(int groupId);
+    List<GroupCollect> selectByGroupId(Long groupId);
 }
